+++ examples/utils.py
@@ -6,7 +6,7 @@ from google.oauth2.credentials import Credentials
 from google_auth_oauthlib.flow import InstalledAppFlow
 
 
-def load_google_credentials(token_file: str, scopes: List[str]):
+def load_google_credentials(secret_file: str, token_file: str, scopes: List[str]):
     """Shows basic usage of the Google Calendar API.
     Prints the start and name of the next 10 events on the user's calendar.
     """
@@ -22,9 +22,9 @@ def load_google_credentials(token_file: str, scopes: List[str]):
             creds.refresh(Request())
         else:
             flow = InstalledAppFlow.from_client_secrets_file(
-                './credentials/google.json', scopes
+                client_secrets_file=secret_file, scopes=scopes
             )
-            creds = flow.run_local_server(port=0)
+            creds = flow.run_local_server(port=19140,redirect_uri_trailing_slash=False)
         # Save the credentials for the next run
         with open(token_file, 'w') as token:
             token.write(creds.to_json())
@@ -32,15 +32,21 @@ def load_google_credentials(token_file: str, scopes: List[str]):
     return creds
 
 
-def load_gmail_credentials():
+def load_gmail_credentials(config_file: str | None = None):
+    if config_file is None:
+        config_file = os.getenv('CONFIG_FILE', ".")
     return load_google_credentials(
-        './credentials/gmail_token.json',
-        ['https://mail.google.com/']
+        secret_file='/'.join([config_file, 'credentials/google.json']),
+        token_file='/'.join([config_file, 'credentials/gmail_token.json']),
+        scopes=['https://mail.google.com/'],
     )
 
 
-def load_google_calendar_credentials():
+def load_google_calendar_credentials(config_file: str | None = None):
+    if config_file is None:
+        config_file = os.getenv('CONFIG_FILE', ".")
     return load_google_credentials(
-        './credentials/gc_token.json',
-        ['https://www.googleapis.com/auth/calendar']
+        secret_file='/'.join([config_file, 'credentials/google.json']),
+        token_file='/'.join([config_file, 'credentials/gc_token.json']),
+        scopes=['https://www.googleapis.com/auth/calendar'],
     )
