+++ npiai/__init__.py
@@ -1,5 +1,4 @@
 from .core import App, create_agent, function
-from .core.browser_app import BrowserApp, NavigatorAgent
 from .llm import OpenAI, LLM
 
-__all__ = ['App', 'create_agent', 'function', 'BrowserApp', 'NavigatorAgent', 'OpenAI', 'LLM']
+__all__ = ['App', 'create_agent', 'function', 'OpenAI', 'LLM']
