+++ npiai/core/__init__.py
@@ -1,20 +1,18 @@
-from npiai.core.app import App, Agent, function
-from npiai.core.browser_app import BrowserApp, BrowserAgent, PlaywrightContext, NavigatorAgent
-from npiai.core.create_agent import create_agent
+from npiai.core.app import App, function
+from npiai.core.app.browser import BrowserApp, PlaywrightContext
+from npiai.core.agent import create_agent, Agent, BrowserAgent, NavigatorAgent
 from npiai.core.hitl import HITL
 from npiai.core.base import BaseAgent, BaseApp, ToolSet
 
 __all__ = [
+    'ToolSet',
+    'function',
     'App',
     'BrowserApp',
-    'function',
     'PlaywrightContext',
-    'create_agent',
     'NavigatorAgent',
     'HITL',
-    'BaseAgent',
-    'BaseApp',
-    'ToolSet',
     'Agent',
-    'BrowserAgent'
+    'BrowserAgent',
+    'create_agent',
 ]
