+++ npiai/core/agent/__init__.py
@@ -1,0 +1,4 @@
+from .agent import create_agent, Agent, BrowserAgent
+from .navigator_agent import NavigatorAgent
+
+__all__ = ['Agent', 'BrowserAgent', 'create_agent', 'NavigatorAgent']
