+++ npiai/core/agent/navigator_agent.py
@@ -10,12 +10,11 @@ from playwright.async_api import Error
 
 from typing_extensions import NotRequired, TypedDict
 
-from npiai import LLM
+from npiai.llm import LLM
 from npiai.utils import logger
 
-from npiai.core.browser_app.browser_app import BrowserApp
-from npiai.core.browser_app.browser_agent import BrowserAgent
-from npiai.core.browser_app.playwright_context import PlaywrightContext
+from npiai.core.app.browser import BrowserApp, PlaywrightContext
+from .agent import BrowserAgent
 
 __PROMPT__ = """
 Imagine that you are imitating humans doing web navigation for a task step by step. I will provide you with the following context and you should call the best tool to fulfill the task.
