+++ npiai/core/app/browser/__init__.py
@@ -1,0 +1,5 @@
+from .app import BrowserApp
+from .playwright_context import PlaywrightContext
+from npiai.core.agent.navigator_agent import NavigatorAgent
+
+__all__ = ['BrowserApp', 'PlaywrightContext', 'NavigatorAgent']
