+++ npiai/core/base.py
@@ -1,6 +1,6 @@
 import os
 from abc import ABC, abstractmethod
-from typing import List
+from typing import Dict, List, Optional, Any
 
 import yaml
 from litellm.types.completion import ChatCompletionToolMessageParam
