--- npiai/core/browser_app/__init__.py
@@ -1,6 +1,0 @@
-from npiai.core.browser_app.browser_app import BrowserApp
-from npiai.core.browser_app.browser_agent import BrowserAgent
-from npiai.core.browser_app.playwright_context import PlaywrightContext
-from npiai.core.browser_app.navigator_agent import NavigatorAgent
-
-__all__ = ['BrowserApp', 'BrowserAgent', 'PlaywrightContext', 'NavigatorAgent']
